- Added x64 builds to all libraries and projects (yay!).  This required significant changes to the project settings - see the Documentation page for some details about what changed. - Added transformation methods to the Entity3D classes. - Fixed a bug in Actor::RemoveElement method - Fixed a bug in Scene::RemoveActor method - Changed accessibility of the CompositeShape class shape container to public.

+++ trunk/Hieroglyph3/Source/Rendering/RendererDX11.cpp
@@ -184,7 +184,7 @@ bool RendererDX11::Initialize( D3D_DRIVER_TYPE DriverType, D3D_FEATURE_LEVEL Fea
 	std::vector<DXGIAdapter*> vAdapters;
 	std::vector<DXGIOutput*> vOutputs;
 
-	while( pFactory->EnumAdapters1( vAdapters.size(), &pCurrentAdapter ) != DXGI_ERROR_NOT_FOUND )
+	while( pFactory->EnumAdapters1( static_cast<UINT>(vAdapters.size()), &pCurrentAdapter ) != DXGI_ERROR_NOT_FOUND )
 	{
 		vAdapters.push_back( new DXGIAdapter( pCurrentAdapter ) );
